Abstract roles guard.
