+++ packages/server/src/course/course-roles.guard.ts
@@ -1,38 +1,18 @@
-import { Injectable, CanActivate, ExecutionContext } from '@nestjs/common';
-import { Reflector } from '@nestjs/core';
+import { Injectable } from '@nestjs/common';
 import { UserModel } from '../profile/user.entity';
+import { RolesGuard } from 'guards/role.guard';
 
 @Injectable()
-export class RolesGuard implements CanActivate {
-  constructor(private reflector: Reflector) {}
-
-  async canActivate(context: ExecutionContext): Promise<boolean> {
-    const roles = this.reflector.get<string[]>('roles', context.getHandler());
-    if (!roles) {
-      return true;
-    }
-    const request = context.switchToHttp().getRequest();
+export class CourseRolesGuard extends RolesGuard {
+  // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
+  async setupData(
+    request: any,
+  ): Promise<{ courseId: number; user: UserModel }> {
     const user = await UserModel.findOne(request.user.userId, {
       relations: ['courses'],
     });
 
     const courseId = request.params.id;
-    return this.matchRoles(roles, user, courseId);
-  }
-
-  matchRoles(roles: string[], user: UserModel, courseId: number): boolean {
-    const userCourse = user.courses.find((course) => {
-      return Number(course.courseId) === Number(courseId);
-    });
-
-    if (!userCourse) {
-      return false;
-    }
-
-    const remaining = roles.filter((role) => {
-      return userCourse.role.toString() === role;
-    });
-
-    return remaining.length > 0;
+    return { courseId, user };
   }
 }
