+++ packages/server/src/question/question.controller.ts
@@ -31,9 +31,10 @@ import { UserModel } from '../profile/user.entity';
 import { QueueModel } from '../queue/queue.entity';
 import { QuestionModel } from './question.entity';
 import { Roles } from 'profile/roles.decorator';
+import { QuestionRolesGuard } from './question-role.guard';
 
 @Controller('questions')
-@UseGuards(JwtAuthGuard)
+@UseGuards(JwtAuthGuard, QuestionRolesGuard)
 @UseInterceptors(ClassSerializerInterceptor)
 export class QuestionController {
   constructor(
